LayoutTests/imported/w3c/web-platform-tests/navigation-api/ordering-and-transition/back-same-document-intercept-reject_no-currententrychange-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: boo

Harness Error (FAIL), message = Unhandled rejection: boo

FAIL event and promise ordering for same-document navigation.back() intercepted by passing a rejected promise to intercept() assert_equals: event 2 (handler run): location.hash value expected "" but got "#1"
